% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrtf.R
\name{unrtf}
\alias{unrtf}
\title{Convert rtf}
\usage{
unrtf(file = NULL, format = c("text", "html", "latex"), verbose = FALSE)
}
\arguments{
\item{file}{path or url to the 'rtf' file}

\item{format}{output format, must be "text", "html" or "latex"}

\item{verbose}{print some output to stderr}
}
\description{
Converts a document in 'rtf' format to other formats
}
