% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrtf.R
\name{unrtf}
\alias{unrtf}
\title{Convert rtf}
\usage{
unrtf(file = NULL, format = c("html", "text", "latex"), verbose = FALSE)
}
\arguments{
\item{file}{path or url to the 'rtf' file}

\item{format}{output format, must be "text", "html" or "latex"}

\item{verbose}{print some output to stderr}
}
\description{
Converts an rtf document to html, text or latex. Note that html output is recommended 
because unrtf has limited support for converting between character encodings.
}
\examples{
library(unrtf)
text <- unrtf("http://www-igm.univ-mlv.fr/~mac/ENS/01-projets/XMLV/pelleas/exemples/Sample.rtf")
cat(text)
}
